src/image/requirements.txt
flask
flask-restx
flask-cors
requests
docker
gitpython